+++ packages/odata-service-inquirer/src/prompts/prompts.ts
@@ -12,9 +12,8 @@ import {
 } from '../types';
 import { getLocalCapProjectPrompts } from './datasources/cap-project/questions';
 import { getMetadataFileQuestion } from './datasources/metadata-file';
-import { getAbapOnPremQuestions } from './datasources/sap-system/abap-on-prem/questions';
-import type { NewSystemAnswers, SystemSelectionAnswer } from './datasources/sap-system/new-system/questions';
-import { newSystemChoiceValue } from './datasources/sap-system/new-system/questions';
+import type { SystemSelectionAnswer } from './datasources/sap-system/new-system/questions';
+import { getNewSystemQuestions, newSystemChoiceValue } from './datasources/sap-system/new-system/questions';
 import { getServiceUrlQuestions } from './datasources/service-url/questions';
 import LoggerHelper from './logger-helper';
 import { getDatasourceTypeChoices } from './prompt-helpers';
@@ -71,7 +70,7 @@ function getDatasourceTypeQuestion(options?: DatasourceTypePromptOptions): YUIQu
             if (source === DatasourceType.businessHub) {
                 return {
                     message: t('prompts.nonUIServiceTypeWarningMessage', {
-                        serviceTypeDesc: t('prompts.datasourceType.businessHubName')
+                        serviceType: t('prompts.datasourceType.businessHubName')
                     }),
                     severity: Severity.warning
                 };
@@ -112,14 +111,12 @@ async function getDatasourceTypeConditionalQuestions(
         ) as OdataServiceQuestion[])
     );
 
-    // Temp integration into Service Inquirer new system questions
     conditionalQuestions.push(
         ...(withCondition(
-            getAbapOnPremQuestions(promptOptions) as Question[],
+            getNewSystemQuestions(promptOptions) as Question[],
             (answers: Answers) =>
                 (answers as OdataServiceAnswers).datasourceType === DatasourceType.sapSystem &&
-                (answers as SystemSelectionAnswer).system === newSystemChoiceValue &&
-                (answers as NewSystemAnswers).newSystemType === 'abapOnPrem'
+                (answers as SystemSelectionAnswer).system === newSystemChoiceValue
         ) as OdataServiceQuestion[])
     );