+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.0.15
+
+### Patch Changes
+
+-   Updated dependencies [bb4305f]
+    -   @sap-ux/fe-fpm-writer@0.28.0
+
+## 0.0.14
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.27.7
+
+## 0.0.13
+
+### Patch Changes
+
+-   Updated dependencies [087752f]
+    -   @sap-ux/fe-fpm-writer@0.27.6
+
 ## 0.0.12
 
 ### Patch Changes